refactor: extract EntryTitle component from EntryContent
